Adds opencv_to_pil helper
