+++ pitop/camera/__init__.py
@@ -1,2 +1,2 @@
 from .camera import Camera  # noqa: F401
-from .pil_to_opencv import pil_to_opencv  # noqa: F401
+from .pil_opencv_conversion import pil_to_opencv, opencv_to_pil  # noqa: F401
